refactor(config-ui): move file components/gitlab/request to components/utils/request
