+++ config-ui/src/components/gitlab/project-selector/use-gitlab-project-selector.ts
@@ -18,7 +18,8 @@
 
 import { useState, useEffect, useMemo } from 'react'
 
-import request from '../request'
+import request from '@/components/utils/request'
+
 import { getGitLabProxyApiPrefix } from '../config'
 
 export type ItemType = {
